Add files via upload
With the rise of the feminist movement and the growing trend of women becoming heads of households, the question begs to be asked, do women also think about retirement as much as men? The theme song of the feminist movement being “Anything you can do I, I can do better, I can do anything better than you . .” asserting that they are equally matched with men in every facet, it’s no secret that, that assertion is not quite true. On average, men make more money than women do when the skills and education are held equal. Given the data set of the pension plans offered by companies in 1989, I wanted to test the theory that men and women are equal, at least when it comes to retirement preparation. Holding all things constant, I expected to see that men and women on average spend about the same amount of time preparing for retirement. What I ended up finding was that on average, men invest in their company’s pension plan approximately 6 years longer than women. Along with that both parties decrease the amount of money they actually invest in the pension plan as the wealth grows. 
There were approximately 194 observations of men and women, 117 women to 77 men , so I thought that the number would be in the woman’s favor.   Of those 194 observations, women were more likely to have completed a bachelor’s degree and men were more likely to have obtained a Master’s degree.  I also found that the amount of time spent in the pension plan was positively skewed.  This could be due to people getting raises and investing in more stock, people could have switched companies a lot and there didn’t have time to invest in the pension, etc. Knowing that women on average do not spend as much time investing in their company’s pension plan, companies need to make it more attractive to women or get better literature out about their pension plane. The more a person pays into their pension, the better off a company should be and the higher likelihood that the person will stay with the company creating low turnover rates.